requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
Pillow==10.0.1
requests==2.31.0
SQLAlchemy==2.0.21
python-dotenv==1.0.0
stripe==6.6.0
paypalrestsdk==1.13.3
psycopg2-binary==2.9.9
gunicorn==21.2.0
modal==0.64.126
email-validator==2.1.0
MarkupSafe==2.1.3
replicate>=1.0.7
bcrypt==4.0.1
mistune==3.1.3
# Force Railway rebuild - 2025-08-22 06:22 AM UTC
